Added support for calculated attributes
Models can now have calculated attributes. A calculated attribute is one whose value is not persisted, and is calculated using values from other attributes and/or relationships on the Model.

This is accomplished by setting a class name and callback method. When the attribute is accessed, this method will be called to generate the value. The method must be set as ```static``` in the owning class, and must handle one argument: ```Model $model```

The YAML metadata driver supports this by setting ```class``` and ```method``` under ```[modelTypeKey].attributes.[attrKey].calculated```