Rely on model's `isDirty()` over `getState()`
The `getState()->is('dirty')` method only updates when changes are made to the owning model directly, and does not recognize changes made on ancestor embedded models. The `isDirty()` check, however, does not have this problem because it actively introspects ancestor embeds via `hasOneEmbeds->areDirty()` and `hasManyEmbeds->areDirty()`.

As such, the `Store` needs to use this method when determining when to update the model, otherwise changes to ancestor embeds may not be committed.